% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\docType{data}
\name{unemp}
\alias{unemp}
\title{USA Unemployment in 2009}
\format{
A data frame with 3218 rows and 8 variables:
\describe{
  \item{id}{the county ID number}
  \item{state_fips}{the state FIPS number}
  \item{county_fips}{the county FIPS number}
  \item{name}{the county name}
  \item{year}{the year}
  \item{rate}{the unemployment rate}
  \item{county}{the county abbreviated name}
  \item{state}{the state acronym}
}
}
\source{
\url{http://datasets.flowingdata.com/unemployment09.csv}
}
\usage{
unemp
}
\description{
A data set containing the 2009 unemployment data in the USA by
 county.
}
\keyword{datasets}
