% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scales.R
\name{viridis_pal}
\alias{viridis_pal}
\title{Viridis palette (discrete)}
\usage{
viridis_pal(alpha = 1)
}
\arguments{
\item{alpha}{pass through parameter to \code{viridis}}
}
\description{
Viridis palette (discrete)
}
\details{
See \link[viridis]{viridis} for more information on the color scale.
}
\examples{
library(scales)
show_col(viridis_pal()(10))
}
\author{
Bob Rudis \email{bob@rudis.net}
}

