% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{viridis}
\alias{inferno}
\alias{magma}
\alias{plasma}
\alias{cividis}
\alias{mako}
\alias{rocket}
\alias{turbo}
\alias{viridis.map}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{viridisLite}{\code{\link[viridisLite:viridis]{cividis}}, \code{\link[viridisLite:viridis]{inferno}}, \code{\link[viridisLite:viridis]{magma}}, \code{\link[viridisLite:viridis]{mako}}, \code{\link[viridisLite:viridis]{plasma}}, \code{\link[viridisLite:viridis]{rocket}}, \code{\link[viridisLite:viridis]{turbo}}, \code{\link[viridisLite]{viridis}}, \code{\link[viridisLite]{viridis.map}}}
}}

