% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\docType{data}
\name{viridis.map}
\alias{viridis.map}
\title{Original 'viridis' color map}
\format{A data frame with 1024 rows and 4 variables:
\itemize{
  \item R: Red value
  \item G: Green value
  \item B: Blue value
  \item opt: The colormap "option" (A: magma; B: inferno; C: plasma; D: viridis)
}}
\usage{
viridis.map
}
\description{
A data.frame containing the original RGB values of the default Matplotlib color
 map ('viridis'). Source: \url{https://github.com/BIDS/colormap/blob/master/option_d.py}.
}
\keyword{datasets}
