% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/viridis.R
\docType{data}
\name{viridis.map}
\alias{viridis.map}
\title{Original 'viridis color map}
\format{A data frame with 256 rows and 3 variables:
\itemize{
  \item R: Red value
  \item G: Green value
  \item B: Blue value
}}
\usage{
viridis.map
}
\description{
A dataset containing the original RGB values of the default Matplotlib color
 map ('viridis'). Source: \url{https://github.com/BIDS/colormap/blob/master/option_d.py}.
}
\keyword{datasets}

